final project/comments_generator/requirements.txt
transformers
bitsandbytes
peft
accelerate
einops
evaluate
scikit-learn
sentencepiece
wandb
dataset
tokenizer